11-multi-agent-routing/pyproject.toml
[project]
name = "my-bot"
version = "0.12.0"
description = "Step 11: Multi-Agent Routing"
requires-python = ">=3.11"
dependencies = [
    "litellm>=1.40.0",
    "anthropic>=0.30.0",
    "rich>=13.0.0",
    "typer>=0.12.0",
    "pyyaml>=6.0",
    "httpx>=0.27.0",
    "pydantic>=2.0",
    "watchdog>=4.0",
    "python-telegram-bot>=21.0",
    "fastapi>=0.110.0",
    "uvicorn[standard]>=0.29.0",
]

[project.scripts]
my-bot = "mybot.cli.main:app"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/mybot"]
